+++ src/main/java/com/yimei/hs/mapper/CangFeeMapper.java
@@ -5,7 +5,7 @@ import com.yimei.hs.entity.CangFee;
 public interface CangFeeMapper {
     int deleteByPrimaryKey(Long id);
 
-    long insert(CangFee record);
+    int insert(CangFee record);
 
     int insertSelective(CangFee record);
