+++ src/main/java/com/yimei/hs/mapper/CangInvoiceDetailMapper.java
@@ -5,7 +5,7 @@ import com.yimei.hs.entity.CangInvoiceDetail;
 public interface CangInvoiceDetailMapper {
     int deleteByPrimaryKey(Long id);
 
-    long insert(CangInvoiceDetail record);
+    int insert(CangInvoiceDetail record);
 
     int insertSelective(CangInvoiceDetail record);
