+++ src/main/java/com/yimei/hs/mapper/CangOrderConfigMapper.java
@@ -5,7 +5,7 @@ import com.yimei.hs.entity.CangOrderConfig;
 public interface CangOrderConfigMapper {
     int deleteByPrimaryKey(Long id);
 
-    long insert(CangOrderConfig record);
+    int insert(CangOrderConfig record);
 
     int insertSelective(CangOrderConfig record);
