+++ src/main/java/com/yimei/hs/mapper/CangSettleTrafficMapper.java
@@ -5,7 +5,7 @@ import com.yimei.hs.entity.CangSettleTraffic;
 public interface CangSettleTrafficMapper {
     int deleteByPrimaryKey(Long id);
 
-    long insert(CangSettleTraffic record);
+    int insert(CangSettleTraffic record);
 
     int insertSelective(CangSettleTraffic record);
