+++ src/main/java/com/yimei/hs/mapper/CangTransferMapper.java
@@ -5,7 +5,7 @@ import com.yimei.hs.entity.CangTransfer;
 public interface CangTransferMapper {
     int deleteByPrimaryKey(Long id);
 
-    long insert(CangTransfer record);
+    int insert(CangTransfer record);
 
     int insertSelective(CangTransfer record);
