+++ src/main/java/com/yimei/hs/mapper/YingFayunMapper.java
@@ -5,7 +5,7 @@ import com.yimei.hs.entity.YingFayun;
 public interface YingFayunMapper {
     int deleteByPrimaryKey(Long id);
 
-    long insert(YingFayun record);
+    int insert(YingFayun record);
 
     int insertSelective(YingFayun record);
