+++ src/main/java/com/yimei/hs/mapper/YingOrderConfigMapper.java
@@ -5,7 +5,7 @@ import com.yimei.hs.entity.YingOrderConfig;
 public interface YingOrderConfigMapper {
     int deleteByPrimaryKey(Long id);
 
-    long insert(YingOrderConfig record);
+    int insert(YingOrderConfig record);
 
     int insertSelective(YingOrderConfig record);
