+++ src/main/java/com/yimei/hs/mapper/YingSettleDownstreamMapper.java
@@ -5,7 +5,7 @@ import com.yimei.hs.entity.YingSettleDownstream;
 public interface YingSettleDownstreamMapper {
     int deleteByPrimaryKey(Long id);
 
-    long insert(YingSettleDownstream record);
+    int insert(YingSettleDownstream record);
 
     int insertSelective(YingSettleDownstream record);
